+++ src/components/orderList/OrderDetail.tsx
@@ -1,13 +1,13 @@
 import { styled } from "styled-components";
 import { useParams } from "react-router-dom";
-import { orderSelector } from "recoil/order";
+import { orderItemsSelector, orderSelector } from "recoil/order";
 import { useRecoilValue } from "recoil";
 import OrderProduct from "./OrderProduct";
 
-const OrderItemDetail = () => {
+const OrderDetail = () => {
   const { id } = useParams();
-  const orderItems = useRecoilValue(orderSelector(Number(id)));
-  const totalPrice = orderItems?.reduce((sum, item) => sum + item.total, 0);
+  const orderItems = useRecoilValue(orderItemsSelector(Number(id)));
+  const order = useRecoilValue(orderSelector(Number(id)));
 
   return (
     <Wrapper>
@@ -27,7 +27,10 @@ const OrderItemDetail = () => {
           </>
         ))}
       </ListBox>
-      <ResultBox>총 결제 금액: {totalPrice?.toLocaleString()}원</ResultBox>
+      <ResultBox>
+        <PriceBox>배송비: {order?.deliveryFee.toLocaleString()}원</PriceBox>
+        <PriceBox>총 결제 금액: {order?.total?.toLocaleString()}원</PriceBox>
+      </ResultBox>
     </Wrapper>
   );
 };
@@ -80,13 +83,18 @@ const ResultBox = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
+  gap: 10px;
 
   height: 100px;
   border: 2px solid rgba(170, 170, 170, 1);
   background: rgba(246, 246, 246, 1);
-  text-align: center;
   font-size: 18px;
   font-weight: 600;
+
+  & > div {
+    text-align: center;
+    width: 100%;
+  }
 `;
 
-export default OrderItemDetail;
+export default OrderDetail;
